extract IncrementalExecutor from Executor
such that IncrementalExecutor extends a base Executor class that ignores the incremental directives. In contrast, the ExecutorThrowingOnIncremental subclass should throw on encountering incremental directives, to be used in the standard `execute` function, signalling that these directives are deprecated unless used intentionally with the new experimental behavior via the new function.
