+++ src/execution/__tests__/cancellation-test.ts
@@ -8,12 +8,10 @@ import { resolveOnNextTick } from '../../__testUtils__/resolveOnNextTick.js';
 import { isAsyncIterable } from '../../jsutils/isAsyncIterable.js';
 import { promiseWithResolvers } from '../../jsutils/promiseWithResolvers.js';
 
-import type { DocumentNode } from '../../language/ast.js';
 import { parse } from '../../language/parser.js';
 
 import {
   GraphQLInterfaceType,
-  GraphQLList,
   GraphQLNonNull,
   GraphQLObjectType,
 } from '../../type/definition.js';
@@ -22,38 +20,7 @@ import { GraphQLSchema } from '../../type/schema.js';
 
 import { buildSchema } from '../../utilities/buildASTSchema.js';
 
-import {
-  execute,
-  experimentalExecuteIncrementally,
-  subscribe,
-} from '../execute.js';
-import type {
-  InitialIncrementalExecutionResult,
-  SubsequentIncrementalExecutionResult,
-} from '../Executor.js';
-
-async function complete(
-  document: DocumentNode,
-  rootValue: unknown,
-  abortSignal: AbortSignal,
-) {
-  const result = await experimentalExecuteIncrementally({
-    schema,
-    document,
-    rootValue,
-    abortSignal,
-  });
-
-  if ('initialResult' in result) {
-    const results: Array<
-      InitialIncrementalExecutionResult | SubsequentIncrementalExecutionResult
-    > = [result.initialResult];
-    for await (const patch of result.subsequentResults) {
-      results.push(patch);
-    }
-    return results;
-  }
-}
+import { execute, subscribe } from '../execute.js';
 
 const schema = buildSchema(`
   type Todo {
@@ -82,45 +49,6 @@ const schema = buildSchema(`
     foo: String
   }
 `);
-
-const streamQuery = new GraphQLObjectType({
-  fields: {
-    scalarList: {
-      type: new GraphQLList(GraphQLString),
-    },
-    slowScalarList: {
-      type: new GraphQLList(GraphQLString),
-    },
-  },
-  name: 'StreamQuery',
-});
-
-const streamSchema = new GraphQLSchema({ query: streamQuery });
-
-const cancelStreamUserType = new GraphQLObjectType({
-  name: 'CancelStreamUser',
-  fields: {
-    id: { type: GraphQLString },
-  },
-});
-
-const cancelStreamTodoType = new GraphQLObjectType({
-  name: 'CancelStreamTodo',
-  fields: {
-    id: { type: GraphQLString },
-    items: { type: new GraphQLList(GraphQLString) },
-    author: { type: cancelStreamUserType },
-  },
-});
-
-const cancelStreamSchema = new GraphQLSchema({
-  query: new GraphQLObjectType({
-    name: 'CancelStreamQuery',
-    fields: {
-      todos: { type: new GraphQLList(cancelStreamTodoType) },
-    },
-  }),
-});
 
 describe('Execute: Cancellation', () => {
   it('should stop the execution when aborted during object field completion', async () => {
@@ -903,678 +831,5 @@ describe('Execute: Cancellation', () => {
       'This operation was aborted',
     );
     expect(returnCalled).to.equal(true);
-  });
-
-  it('should allow deferred execution when not aborted', async () => {
-    const abortController = new AbortController();
-    const document = parse(`
-      query {
-        todo {
-          id
-          ... on Todo @defer {
-            author {
-              id
-            }
-          }
-        }
-      }
-    `);
-
-    const result = await experimentalExecuteIncrementally({
-      document,
-      schema,
-      rootValue: {
-        todo: {
-          author: { id: '1' },
-        },
-      },
-      abortSignal: abortController.signal,
-    });
-
-    assert('initialResult' in result);
-
-    const { initialResult, subsequentResults } = result;
-
-    expectJSON(initialResult).toDeepEqual({
-      data: {
-        todo: {
-          id: null,
-        },
-      },
-      pending: [{ id: '0', path: ['todo'] }],
-      hasNext: true,
-    });
-
-    const payload1 = await subsequentResults.next();
-    expectJSON(payload1).toDeepEqual({
-      done: false,
-      value: {
-        incremental: [
-          {
-            data: { author: { id: '1' } },
-            id: '0',
-          },
-        ],
-        completed: [{ id: '0' }],
-        hasNext: false,
-      },
-    });
-  });
-
-  it('should stop deferred execution when aborted', async () => {
-    const abortController = new AbortController();
-    const document = parse(`
-      query {
-        todo {
-          id
-          ... on Todo @defer {
-            author {
-              id
-            }
-          }
-        }
-      }
-    `);
-
-    const resultPromise = execute({
-      document,
-      schema,
-      rootValue: {
-        todo: async () =>
-          Promise.resolve({
-            id: '1',
-            /* c8 ignore next */
-            author: () => expect.fail('Should not be called'),
-          }),
-      },
-      abortSignal: abortController.signal,
-    });
-
-    abortController.abort();
-
-    await expectPromise(resultPromise).toRejectWith(
-      'This operation was aborted',
-    );
-  });
-
-  it('should stop deferred execution when aborted mid-execution', async () => {
-    const abortController = new AbortController();
-    const document = parse(`
-      query {
-        ... on Query @defer {
-          todo {
-            id
-            author {
-              id
-            }
-          }
-        }
-      }
-    `);
-
-    const resultPromise = complete(
-      document,
-      {
-        todo: () =>
-          Promise.resolve({
-            id: '1',
-            /* c8 ignore next 2 */
-            author: () =>
-              Promise.resolve(() => expect.fail('Should not be called')),
-          }),
-      },
-      abortController.signal,
-    );
-
-    abortController.abort();
-
-    await expectPromise(resultPromise).toRejectWith(
-      'This operation was aborted',
-    );
-  });
-
-  it('should stop streamed execution when aborted', async () => {
-    const abortController = new AbortController();
-    const document = parse(`
-      query {
-        todo {
-          id
-          items @stream
-        }
-      }
-    `);
-
-    const resultPromise = complete(
-      document,
-      {
-        todo: {
-          id: '1',
-          items: [Promise.resolve('item')],
-        },
-      },
-      abortController.signal,
-    );
-
-    abortController.abort();
-
-    await expectPromise(resultPromise).toRejectWith(
-      'This operation was aborted',
-    );
-  });
-
-  it('cancels streaming when aborted during async iterator next', async () => {
-    const abortController = new AbortController();
-    const document = parse('{ scalarList @stream(initialCount: 0) }');
-
-    const { promise: nextStarted, resolve: resolveNextStarted } =
-      // eslint-disable-next-line @typescript-eslint/no-invalid-void-type
-      promiseWithResolvers<void>();
-    const { promise: nextReturned, resolve: resolveNextReturned } =
-      promiseWithResolvers<IteratorResult<unknown>>();
-    let done = false;
-    const asyncIterator = {
-      [Symbol.asyncIterator]() {
-        return this;
-      },
-      next() {
-        if (done) {
-          return Promise.resolve({ value: undefined, done: true });
-        }
-        done = true;
-        resolveNextStarted();
-        return nextReturned;
-      },
-    };
-
-    const result = await experimentalExecuteIncrementally({
-      schema: streamSchema,
-      document,
-      rootValue: {
-        scalarList: () => asyncIterator,
-      },
-      abortSignal: abortController.signal,
-    });
-    assert('initialResult' in result);
-
-    const iterator = result.subsequentResults[Symbol.asyncIterator]();
-    const nextPromise = iterator.next();
-    await nextStarted;
-    abortController.abort();
-
-    resolveNextReturned({ value: 'value', done: false });
-    await expectPromise(nextPromise).toRejectWith('This operation was aborted');
-  });
-
-  it('cancels streaming when aborted while item promise is pending', async () => {
-    const abortController = new AbortController();
-    const document = parse('{ scalarList @stream(initialCount: 0) }');
-
-    const { promise: itemPromise, resolve: resolveItem } =
-      promiseWithResolvers<string>();
-    const { promise: nextStarted, resolve: resolveNextStarted } =
-      // eslint-disable-next-line @typescript-eslint/no-invalid-void-type
-      promiseWithResolvers<void>();
-    let done = false;
-    const asyncIterator = {
-      [Symbol.asyncIterator]() {
-        return this;
-      },
-      next() {
-        if (done) {
-          return Promise.resolve({ value: undefined, done: true });
-        }
-        done = true;
-        resolveNextStarted();
-        return Promise.resolve({ value: itemPromise, done: false });
-      },
-    };
-
-    const result = await experimentalExecuteIncrementally({
-      schema: streamSchema,
-      document,
-      rootValue: {
-        scalarList: () => asyncIterator,
-      },
-      abortSignal: abortController.signal,
-    });
-    assert('initialResult' in result);
-
-    const iterator = result.subsequentResults[Symbol.asyncIterator]();
-    const nextPromise = iterator.next();
-    await nextStarted;
-    abortController.abort();
-
-    resolveItem('value');
-    await expectPromise(nextPromise).toRejectWith('This operation was aborted');
-  });
-
-  it('stops when the stream queue is back-pressured and the consumer cancels', async () => {
-    const document = parse('{ scalarList @stream(initialCount: 0) }');
-    const { promise: reachedCapacity, resolve: resolveReachedCapacity } =
-      // eslint-disable-next-line @typescript-eslint/no-invalid-void-type
-      promiseWithResolvers<void>();
-    let count = 0;
-    let done = false;
-    const iterator = {
-      [Symbol.iterator]() {
-        return this;
-      },
-      next() {
-        if (done) {
-          return { value: undefined, done: true };
-        }
-        count += 1;
-        if (count === 100) {
-          resolveReachedCapacity();
-        }
-        if (count > 100) {
-          done = true;
-        }
-        return { value: String(count), done: false };
-      },
-    };
-
-    const result = await experimentalExecuteIncrementally({
-      schema: streamSchema,
-      document,
-      rootValue: {
-        scalarList: () => iterator,
-      },
-      enableEarlyExecution: true,
-    });
-    assert('initialResult' in result);
-
-    await reachedCapacity;
-    await resolveOnNextTick();
-    const stream = result.subsequentResults[Symbol.asyncIterator]();
-    await expectPromise(stream.return()).toResolve();
-  });
-
-  it('cancels pending deferred execution groups', async () => {
-    const abortController = new AbortController();
-    const { promise: slowPromise } = promiseWithResolvers<unknown>();
-    const document = parse('{ scalarList ... @defer { slowScalarList } }');
-
-    const result = await experimentalExecuteIncrementally({
-      schema: streamSchema,
-      document,
-      rootValue: {
-        scalarList: () => ['a'],
-        slowScalarList: () => slowPromise,
-      },
-      enableEarlyExecution: true,
-      abortSignal: abortController.signal,
-    });
-    assert('initialResult' in result);
-
-    const iterator = result.subsequentResults[Symbol.asyncIterator]();
-    abortController.abort();
-
-    await expectPromise(iterator.next()).toRejectWith(
-      'This operation was aborted',
-    );
-  });
-
-  it('cancels tasks and streams when aborted before initial execution finishes', async () => {
-    const abortController = new AbortController();
-    const document = parse(`
-      query {
-        todo {
-          id
-          items @stream(initialCount: 0)
-          ... @defer {
-            author {
-              id
-            }
-          }
-        }
-        blocker
-      }
-    `);
-
-    const { promise: blockerPromise, resolve: resolveBlocker } =
-      promiseWithResolvers<string>();
-    const { promise: blockerStarted, resolve: resolveBlockerStarted } =
-      // eslint-disable-next-line @typescript-eslint/no-invalid-void-type
-      promiseWithResolvers<void>();
-    const { promise: itemsStarted, resolve: resolveItemsStarted } =
-      // eslint-disable-next-line @typescript-eslint/no-invalid-void-type
-      promiseWithResolvers<void>();
-
-    const resultPromise = experimentalExecuteIncrementally({
-      schema,
-      document,
-      abortSignal: abortController.signal,
-      rootValue: {
-        blocker() {
-          resolveBlockerStarted();
-          return blockerPromise;
-        },
-        todo: {
-          id: 'todo',
-          items() {
-            resolveItemsStarted();
-            return ['a', 'b'];
-          },
-          author() {
-            return { id: 'author' };
-          },
-        },
-      },
-    });
-
-    await itemsStarted;
-    await blockerStarted;
-
-    abortController.abort();
-
-    await expectPromise(resultPromise).toRejectWith(
-      'This operation was aborted',
-    );
-
-    resolveBlocker('done');
-  });
-
-  it('should ignore repeated cancellation attempts during incremental execution', async () => {
-    const abortController = new AbortController();
-    const document = parse(`
-      query {
-        todo {
-          id
-          items @stream(initialCount: 0)
-          ... @defer {
-            author {
-              id
-            }
-          }
-        }
-      }
-    `);
-
-    let streamReturnCount = 0;
-    let nextPromiseResolved = false;
-    const { promise: nextStarted, resolve: resolveNextStarted } =
-      // eslint-disable-next-line @typescript-eslint/no-invalid-void-type
-      promiseWithResolvers<void>();
-    const { promise: nextPromise, resolve: resolveNext } =
-      promiseWithResolvers<IteratorResult<string>>();
-    const asyncIterator = {
-      [Symbol.asyncIterator]() {
-        return this;
-      },
-      next() {
-        if (nextPromiseResolved) {
-          return Promise.resolve({ value: undefined, done: true });
-        }
-        nextPromiseResolved = true;
-        resolveNextStarted();
-        return nextPromise;
-      },
-      return() {
-        streamReturnCount += 1;
-        return Promise.resolve({ value: undefined, done: true });
-      },
-    };
-
-    const { promise: authorStarted, resolve: resolveAuthorStarted } =
-      // eslint-disable-next-line @typescript-eslint/no-invalid-void-type
-      promiseWithResolvers<void>();
-    const { promise: authorPromise, resolve: resolveAuthor } =
-      promiseWithResolvers<{ id: string }>();
-    const rootValue = {
-      todo: {
-        id: 'todo',
-        items: () => asyncIterator,
-        author() {
-          resolveAuthorStarted();
-          return authorPromise;
-        },
-      },
-    };
-
-    const result = await experimentalExecuteIncrementally({
-      schema,
-      document,
-      rootValue,
-      enableEarlyExecution: true,
-      abortSignal: abortController.signal,
-    });
-    assert('initialResult' in result);
-
-    const iterator = result.subsequentResults[Symbol.asyncIterator]();
-    const nextResultPromise = iterator.next();
-
-    await authorStarted;
-    await nextStarted;
-
-    abortController.abort();
-    await resolveOnNextTick();
-
-    let firstResult:
-      | IteratorResult<SubsequentIncrementalExecutionResult>
-      | undefined;
-    try {
-      firstResult =
-        (await nextResultPromise) as IteratorResult<SubsequentIncrementalExecutionResult>;
-    } catch (error) {
-      expect(error).to.be.instanceOf(Error);
-      expect((error as Error).message).to.equal('This operation was aborted');
-    }
-    if (firstResult && !firstResult.done) {
-      try {
-        const followUp = await iterator.next();
-        expect(followUp.done).to.equal(true);
-      } catch (error) {
-        expect(error).to.be.instanceOf(Error);
-        expect((error as Error).message).to.equal('This operation was aborted');
-      }
-    }
-    expect(streamReturnCount).to.equal(1);
-
-    const priorStreamReturnCount = streamReturnCount;
-    abortController.abort();
-    await resolveOnNextTick();
-
-    expect(streamReturnCount).to.equal(priorStreamReturnCount);
-
-    resolveNext({ value: 'value', done: false });
-    resolveAuthor({ id: 'author' });
-  });
-
-  it('should ignore deferred payloads resolved after cancellation', async () => {
-    const abortController = new AbortController();
-    const document = parse(`
-      query {
-        todo {
-          id
-          ... @defer {
-            author {
-              id
-            }
-          }
-        }
-      }
-    `);
-
-    const { promise: authorStarted, resolve: resolveAuthorStarted } =
-      // eslint-disable-next-line @typescript-eslint/no-invalid-void-type
-      promiseWithResolvers<void>();
-    const { promise: authorPromise, resolve: resolveAuthor } =
-      promiseWithResolvers<{ id: string }>();
-
-    const result = await experimentalExecuteIncrementally({
-      schema,
-      document,
-      abortSignal: abortController.signal,
-      enableEarlyExecution: true,
-      rootValue: {
-        todo: {
-          id: 'todo',
-          author() {
-            resolveAuthorStarted();
-            return authorPromise;
-          },
-        },
-      },
-    });
-    assert('initialResult' in result);
-
-    const iterator = result.subsequentResults[Symbol.asyncIterator]();
-    const nextResultPromise = iterator.next();
-
-    await authorStarted;
-    abortController.abort();
-
-    resolveAuthor({ id: 'author' });
-
-    await expectPromise(nextResultPromise).toRejectWith(
-      'This operation was aborted',
-    );
-    await expectPromise(authorPromise).toResolve();
-  });
-
-  it('should ignore deferred errors after cancellation', async () => {
-    const abortController = new AbortController();
-    const document = parse(`
-      query {
-        todo {
-          id
-          ... @defer {
-            author {
-              id
-            }
-          }
-        }
-      }
-    `);
-
-    const { promise: authorStarted, resolve: resolveAuthorStarted } =
-      // eslint-disable-next-line @typescript-eslint/no-invalid-void-type
-      promiseWithResolvers<void>();
-    const { promise: authorPromise, reject: rejectAuthor } =
-      promiseWithResolvers<{ id: string }>();
-
-    const result = await experimentalExecuteIncrementally({
-      schema,
-      document,
-      abortSignal: abortController.signal,
-      enableEarlyExecution: true,
-      rootValue: {
-        todo: {
-          id: 'todo',
-          author() {
-            resolveAuthorStarted();
-            return authorPromise;
-          },
-        },
-      },
-    });
-    assert('initialResult' in result);
-
-    const iterator = result.subsequentResults[Symbol.asyncIterator]();
-    const nextResultPromise = iterator.next();
-
-    await authorStarted;
-    abortController.abort();
-
-    rejectAuthor(new Error('late error'));
-
-    await expectPromise(nextResultPromise).toRejectWith(
-      'This operation was aborted',
-    );
-    await expectPromise(authorPromise).toRejectWith('late error');
-  });
-
-  it('cancels stream item executors with deferred work and nested streams', async () => {
-    const document = parse(`
-      query {
-        todos @stream(initialCount: 0) {
-          id
-          items @stream(initialCount: 0)
-          ... @defer {
-            author {
-              id
-            }
-          }
-        }
-      }
-    `);
-
-    const { promise: idPromise, resolve: resolveId } =
-      promiseWithResolvers<string>();
-    const { promise: idStarted, resolve: resolveIdStarted } =
-      // eslint-disable-next-line @typescript-eslint/no-invalid-void-type
-      promiseWithResolvers<void>();
-
-    const result = await experimentalExecuteIncrementally({
-      schema: cancelStreamSchema,
-      document,
-      enableEarlyExecution: true,
-      rootValue: {
-        todos: [
-          {
-            id() {
-              resolveIdStarted();
-              return idPromise;
-            },
-            items: ['a'],
-            author: { id: 'author' },
-          },
-        ],
-      },
-    });
-    assert('initialResult' in result);
-
-    const iterator = result.subsequentResults[Symbol.asyncIterator]();
-    await idStarted;
-    await expectPromise(iterator.return()).toResolve();
-    await resolveOnNextTick();
-
-    resolveId('todo');
-    await resolveOnNextTick();
-  });
-
-  it('stops streaming when a pending stream item resolves after cancellation', async () => {
-    const document = parse('{ scalarList @stream(initialCount: 0) }');
-    const { promise: itemPromise, resolve: resolveItem } =
-      promiseWithResolvers<string>();
-    const { promise: nextStarted, resolve: resolveNextStarted } =
-      // eslint-disable-next-line @typescript-eslint/no-invalid-void-type
-      promiseWithResolvers<void>();
-    let done = false;
-    const iterator = {
-      [Symbol.iterator]() {
-        return this;
-      },
-      next() {
-        if (done) {
-          return { value: undefined, done: true };
-        }
-        done = true;
-        resolveNextStarted();
-        return { value: itemPromise, done: false };
-      },
-    };
-
-    const result = await experimentalExecuteIncrementally({
-      schema: streamSchema,
-      document,
-      rootValue: {
-        scalarList: () => iterator,
-      },
-    });
-    assert('initialResult' in result);
-
-    const stream = result.subsequentResults[Symbol.asyncIterator]();
-    const nextPromise = stream.next();
-
-    await nextStarted;
-    const returnPromise = stream.return();
-    await resolveOnNextTick();
-
-    resolveItem('value');
-
-    await expectPromise(returnPromise).toResolve();
-    await expectPromise(nextPromise).toResolve();
   });
 });
