+++ src/execution/incremental/__tests__/defer-test.ts
@@ -1,28 +1,31 @@
 import { assert, expect } from 'chai';
 import { describe, it } from 'mocha';
 
-import { expectJSON } from '../../__testUtils__/expectJSON.js';
-import { expectPromise } from '../../__testUtils__/expectPromise.js';
-import { resolveOnNextTick } from '../../__testUtils__/resolveOnNextTick.js';
+import { expectJSON } from '../../../__testUtils__/expectJSON.js';
+import { expectPromise } from '../../../__testUtils__/expectPromise.js';
+import { resolveOnNextTick } from '../../../__testUtils__/resolveOnNextTick.js';
 
-import { promiseWithResolvers } from '../../jsutils/promiseWithResolvers.js';
+import { promiseWithResolvers } from '../../../jsutils/promiseWithResolvers.js';
 
-import type { DocumentNode } from '../../language/ast.js';
-import { parse } from '../../language/parser.js';
+import type { DocumentNode } from '../../../language/ast.js';
+import { parse } from '../../../language/parser.js';
 
 import {
   GraphQLList,
   GraphQLNonNull,
   GraphQLObjectType,
-} from '../../type/definition.js';
-import { GraphQLID, GraphQLString } from '../../type/scalars.js';
-import { GraphQLSchema } from '../../type/schema.js';
+} from '../../../type/definition.js';
+import { GraphQLID, GraphQLString } from '../../../type/scalars.js';
+import { GraphQLSchema } from '../../../type/schema.js';
 
-import { execute, experimentalExecuteIncrementally } from '../execute.js';
+import { buildSchema } from '../../../utilities/buildASTSchema.js';
+
+import { experimentalExecuteIncrementally } from '../../execute.js';
+
 import type {
   InitialIncrementalExecutionResult,
   SubsequentIncrementalExecutionResult,
-} from '../Executor.js';
+} from '../IncrementalExecutor.js';
 
 const friendType = new GraphQLObjectType({
   fields: {
@@ -138,6 +141,36 @@ const query = new GraphQLObjectType({
 
 const schema = new GraphQLSchema({ query });
 
+const cancellationSchema = buildSchema(`
+  type Todo {
+    id: ID
+    items: [String]
+    author: User
+  }
+
+  type User {
+    id: ID
+    name: String
+  }
+
+  type Query {
+    todo: Todo
+    nonNullableTodo: Todo!
+    blocker: String
+    scalarList: [String]
+    slowScalarList: [String]
+  }
+
+  type Mutation {
+    foo: String
+    bar: String
+  }
+
+  type Subscription {
+    foo: String
+  }
+`);
+
 async function complete(
   document: DocumentNode,
   rootValue: unknown = { hero },
@@ -148,6 +181,32 @@ async function complete(
     document,
     rootValue,
     enableEarlyExecution,
+  });
+
+  if ('initialResult' in result) {
+    const results: Array<
+      InitialIncrementalExecutionResult | SubsequentIncrementalExecutionResult
+    > = [result.initialResult];
+    for await (const patch of result.subsequentResults) {
+      results.push(patch);
+    }
+    return results;
+  }
+  return result;
+}
+
+async function completeCancellation(
+  document: DocumentNode,
+  rootValue: unknown,
+  abortSignal: AbortSignal,
+  enableEarlyExecution = false,
+) {
+  const result = await experimentalExecuteIncrementally({
+    schema: cancellationSchema,
+    document,
+    rootValue,
+    enableEarlyExecution,
+    abortSignal,
   });
 
   if ('initialResult' in result) {
@@ -2741,46 +2800,260 @@ describe('Execute: defer directive', () => {
     });
   });
 
-  it('original execute function throws error if anything is deferred and everything else is sync', () => {
-    const doc = `
-      query Deferred {
-        ... @defer { hero { id } }
+  it('should allow deferred execution when passed abortSignal, if not aborted', async () => {
+    const abortController = new AbortController();
+    const document = parse(`
+      query {
+        todo {
+          id
+          ... on Todo @defer {
+            author {
+              id
+            }
+          }
+        }
       }
-    `;
-    expect(() =>
-      execute({
-        schema,
-        document: parse(doc),
-        rootValue: {},
-      }),
-    ).to.throw(
-      'Executing this GraphQL operation would unexpectedly produce multiple payloads (due to @defer or @stream directive)',
+    `);
+
+    const result = await experimentalExecuteIncrementally({
+      schema: cancellationSchema,
+      document,
+      rootValue: {
+        todo: {
+          author: { id: '1' },
+        },
+      },
+      abortSignal: abortController.signal,
+    });
+
+    assert('initialResult' in result);
+
+    const { initialResult, subsequentResults } = result;
+
+    expectJSON(initialResult).toDeepEqual({
+      data: {
+        todo: {
+          id: null,
+        },
+      },
+      pending: [{ id: '0', path: ['todo'] }],
+      hasNext: true,
+    });
+
+    const payload1 = await subsequentResults.next();
+    expectJSON(payload1).toDeepEqual({
+      done: false,
+      value: {
+        incremental: [
+          {
+            data: { author: { id: '1' } },
+            id: '0',
+          },
+        ],
+        completed: [{ id: '0' }],
+        hasNext: false,
+      },
+    });
+  });
+
+  it('should stop deferred execution when aborted', async () => {
+    const abortController = new AbortController();
+    const document = parse(`
+      query {
+        todo {
+          id
+          ... on Todo @defer {
+            author {
+              id
+            }
+          }
+        }
+      }
+    `);
+
+    const resultPromise = experimentalExecuteIncrementally({
+      schema: cancellationSchema,
+      document,
+      rootValue: {
+        todo: async () =>
+          Promise.resolve({
+            id: '1',
+            /* c8 ignore next */
+            author: () => expect.fail('Should not be called'),
+          }),
+      },
+      abortSignal: abortController.signal,
+    });
+
+    abortController.abort();
+
+    await expectPromise(resultPromise).toRejectWith(
+      'This operation was aborted',
     );
   });
 
-  it('original execute function resolves to error if anything is deferred and something else is async', async () => {
-    const doc = `
-      query Deferred {
-        hero { name }
-        ... @defer { hero { id } }
+  it('should stop deferred execution when aborted mid-execution', async () => {
+    const abortController = new AbortController();
+    const document = parse(`
+      query {
+        ... on Query @defer {
+          todo {
+            id
+            author {
+              id
+            }
+          }
+        }
       }
-    `;
-    await expectPromise(
-      execute({
-        schema,
-        document: parse(doc),
-        rootValue: {
-          hero: {
-            ...hero,
-            name: async () => {
-              await resolveOnNextTick();
-              return 'slow';
-            },
+    `);
+
+    const resultPromise = completeCancellation(
+      document,
+      {
+        todo: () =>
+          Promise.resolve({
+            id: '1',
+            /* c8 ignore next 2 */
+            author: () =>
+              Promise.resolve(() => expect.fail('Should not be called')),
+          }),
+      },
+      abortController.signal,
+    );
+
+    abortController.abort();
+
+    await expectPromise(resultPromise).toRejectWith(
+      'This operation was aborted',
+    );
+  });
+
+  it('cancels pending deferred execution groups', async () => {
+    const abortController = new AbortController();
+    const { promise: slowPromise } = promiseWithResolvers<unknown>();
+    const document = parse('{ scalarList ... @defer { slowScalarList } }');
+
+    const result = await experimentalExecuteIncrementally({
+      schema: cancellationSchema,
+      document,
+      rootValue: {
+        scalarList: () => ['a'],
+        slowScalarList: () => slowPromise,
+      },
+      enableEarlyExecution: true,
+      abortSignal: abortController.signal,
+    });
+    assert('initialResult' in result);
+
+    const iterator = result.subsequentResults[Symbol.asyncIterator]();
+    abortController.abort();
+
+    await expectPromise(iterator.next()).toRejectWith(
+      'This operation was aborted',
+    );
+  });
+
+  it('should ignore deferred payloads resolved after cancellation', async () => {
+    const abortController = new AbortController();
+    const document = parse(`
+      query {
+        todo {
+          id
+          ... @defer {
+            author {
+              id
+            }
+          }
+        }
+      }
+    `);
+
+    const { promise: authorStarted, resolve: resolveAuthorStarted } =
+      // eslint-disable-next-line @typescript-eslint/no-invalid-void-type
+      promiseWithResolvers<void>();
+    const { promise: authorPromise, resolve: resolveAuthor } =
+      promiseWithResolvers<{ id: string }>();
+
+    const result = await experimentalExecuteIncrementally({
+      schema: cancellationSchema,
+      document,
+      abortSignal: abortController.signal,
+      enableEarlyExecution: true,
+      rootValue: {
+        todo: {
+          id: 'todo',
+          author() {
+            resolveAuthorStarted();
+            return authorPromise;
           },
         },
-      }),
-    ).toRejectWith(
-      'Executing this GraphQL operation would unexpectedly produce multiple payloads (due to @defer or @stream directive)',
+      },
+    });
+    assert('initialResult' in result);
+
+    const iterator = result.subsequentResults[Symbol.asyncIterator]();
+    const nextResultPromise = iterator.next();
+
+    await authorStarted;
+    abortController.abort();
+
+    resolveAuthor({ id: 'author' });
+
+    await expectPromise(nextResultPromise).toRejectWith(
+      'This operation was aborted',
     );
+    await expectPromise(authorPromise).toResolve();
+  });
+
+  it('should ignore deferred errors after cancellation', async () => {
+    const abortController = new AbortController();
+    const document = parse(`
+      query {
+        todo {
+          id
+          ... @defer {
+            author {
+              id
+            }
+          }
+        }
+      }
+    `);
+
+    const { promise: authorStarted, resolve: resolveAuthorStarted } =
+      // eslint-disable-next-line @typescript-eslint/no-invalid-void-type
+      promiseWithResolvers<void>();
+    const { promise: authorPromise, reject: rejectAuthor } =
+      promiseWithResolvers<{ id: string }>();
+
+    const result = await experimentalExecuteIncrementally({
+      schema: cancellationSchema,
+      document,
+      abortSignal: abortController.signal,
+      enableEarlyExecution: true,
+      rootValue: {
+        todo: {
+          id: 'todo',
+          author() {
+            resolveAuthorStarted();
+            return authorPromise;
+          },
+        },
+      },
+    });
+    assert('initialResult' in result);
+
+    const iterator = result.subsequentResults[Symbol.asyncIterator]();
+    const nextResultPromise = iterator.next();
+
+    await authorStarted;
+    abortController.abort();
+
+    rejectAuthor(new Error('late error'));
+
+    await expectPromise(nextResultPromise).toRejectWith(
+      'This operation was aborted',
+    );
+    await expectPromise(authorPromise).toRejectWith('late error');
   });
 });
